WIP: Changed Feedback Class structure
The GPU controller is now templated on feedback state (think of it
like Params for the GPU controller) and DDP has been rewritten to
reflect that. There are a lot of things commented out of DDP that
can probably be removed now as well as an extra struct in ddp.cuh
that can be removed (DDPFBState).

Work to be done:
- check for CUDA memleaks on current feedback implementation (should
  be cleaner now as I am just copying a struct to the GPU like in
  Dynamics and Costs)
- rewrite RMPPI kernel to be templated on GPU Feedback controller
- rewrite RMPPI to use above kernel
- rewrite all MPPI, Tube-MPPI, RMPPI tests to make use of the new
template layout and constructor
- create a branch for MPPI-ROS that fixes MPPI, Tube, and RMPPI
construction as well.
- write documentation on how to create a new feedback controller