Adjust colored samples to be smoother
Fix colored samples to smoothly transition at optimization stride point.
This is done by subtracting the control value at optimization stride
point from the entire trajectory. Since this is done in the  noise
section, when the mean is added in the rollout kernel, this ensures a
smooth transition from the previous control which we cannot change to
the new sampled trajectory