WIP: Fix cost kernel edge cases
- Added checks for whether the current num_rollouts/num_timesteps and
block sizes breaks assumptions made in the dynamics or cost kernels
- found a small bug in the rolloutCostKernels sum reduction that would
cause it to not work for block sizes that were not 1, 2, 8, 16, 32, 64.
Now the code will use the faster unrolled summations if the block size
is a power of 2 and will use the slower unroll method otherwise. There
might be a weird edge case with odd multiples of 32 such as 96... Will
need to investigate further
- made better comments on the alignment checking methods in math_utils.h
- added the dynamics as a requirement to the RMPPI Cost Kernel so that
we can properly calculate the feedback gain from an output rather than a
state