Improve colored sampling distributions
The time horizon used for sampling trajectories in the frequency domain was doubled to prevent the control samples from collapsing back to the same starting point in the time domain. This occurs due to the choice of using Hermitian sequences with the iFFT to produce real only components in the time domain. A decay rate to the offset subtraction was also introduced as subtracted the same offset throughout an entire sampling trajectory makes it much more likely to hit control constraints