Fix sampling issues
There was a bug in the GaussianKernel that made it so we sampled pure
noise we sampled pure noise 99% the time instead of 1% when the
control_dim was divisible by 2. Also fixed a bug in the colored sampling
FFT planning that didn't account for CONTROL_DIMs larger than 1.
Finally, fixed some incorrect variable names in the sampling
distribution params.