WIP: removing barriers from cost kernel
We have been occasionally seeing weird shared memory issues in the cost
function and it looks to be due to the combination of coalesced memory
accesses and the use of cuda barriers in the cost kernel. Both should
not be enabled but more testing will need to be done as to which has the
largest benefit to opimization times.