Implement combind rollout kernel again
Comparing MPPI-Generic to the autorally codebase, there are times where
the combined rollout kernel is faster than the split kernels for the
dynamics and cost. This looks to be caused by the split kernels requring
to write the output trajectory to global memory in the dynamics and read
from global memory in the cost kernel.