Start unit tests for kernels
I started to write unit tests for the R-MPPI kernels but discovered
there was a discrepency between the coalesced and non-coalesced reading
of the output in the cost kernel. This led me to the following changes:

- Simplify running cost calculation. We now calculate costs from t = 0
to t = num_timesteps -1 (inclusive of both values), and divide by
num_timesteps. We used to skip t = 0 as in the old code, the value at t
= 0 used to be the initial state and the same for each trajectory. Now,
the output that is stored at t = 0 is the output after our first call to
step() (aka the output at t = 1).
- Make adjustments to how we read outputs in the non-coalesced case in
split cost kernels to reflect the change
- Put templating for coalesced cost function memory reading in the
launch methods so that it can be turned on or off in different calls to
the launch methods.
- Begin rewriting the R-MPPI kernel tests
  - Wrote CPU version of the initEvalKernel
  - Wrote comparison tests of the CPU initEvalKernel to the combined
kernel, the split kernel with no coalesced memory reads, and to the
split kernel with coalesced memory reads
- Still to be done on the R-MPPI Kernel testing side:
  - Write R-MPPI Rollout comparison GPU vs CPU
  - Compare R-MPPI Rollout to standard MPPI rollout by setting the real
and nominal systems to have the same initial state and use the same
noise for sampling controls