Get all unit tests compiling
Rewrote tests for rolloutKernel and dropped comparisons to the Autorally
rolloutKernel since we are now pretty different with new sampling
distribution API and using outputs for the cost function. The
rolloutKernel tests now compare to a CPU implementation and ensure that
the combined kernel, the split kernel with coalesced memory accesses in
the cost function, and the split kernel without all match.

Failure Cases:
- Still need to fix viz_kernel tests
- Failing some dynamics/nn unit tests
- Failing Quadrotor Map cost tests
- Plant tests are segfaulting
- Still occasionally failing automatic kernel selection tests
- Racer Suspension unit tests cause infinite loop of some sort
- Colored Noise tests are currently disabled
- Need unit tests for 2d and 3d texture helpers queryTexture CPU method