WIP: Add checks for partially empty texture maps
When the elevation or normal map is not fully known, it will have NANs
in the unknown regions. I added  isfinite() checks on the texture values
in RacerDubinsElevationSuspension.

Other changes:
- Added unit tests for partial elevation and normal maps for
RacerDubinsElevationSuspension
- Changed RacerDubinsElevation to output NAN wheel forces to indicate
they are not filled in.
- Added a stateFromMap() implementation for
RacerDubinsElevationSuspension to get the new states of POS_Z, VEL_Z,
ROLL_RATE, PITCH_RATE from the buffered plant.